Widened VBO/PBO buffer offsets and sizes to long to better match the GLsizeiptr and GLintptr native types

+++ src/native/common/common_tools.h
@@ -105,7 +105,7 @@ static inline jobject safeNewBufferCached(JNIEnv *env, void *p, int size, jobjec
 
 }
 
-static inline void *offsetToPointer(jint offset) {
+static inline void *offsetToPointer(jlong offset) {
 	return (char *)NULL + offset;
 }
